refactor: move page files to parent directories
