+++ src/pages/gaps/index.tsx
@@ -1,25 +1,25 @@
 import { useContext, useEffect, useState } from 'react'
-import { PageContainer } from './components/PageContainer'
-import { Row } from './components/Row'
-import { Label } from './components/Label'
-import { TEXTS } from '../resources/texts'
-import OperatorSelector from './components/OperatorSelector'
-import LineNumberSelector from './components/LineSelector'
-import { SearchContext } from '../model/pageState'
-import { Gap, GapsList } from '../model/gaps'
-import { getGapsAsync } from '../api/gapsService'
+import { PageContainer } from '../components/PageContainer'
+import { Row } from '../components/Row'
+import { Label } from '../components/Label'
+import { TEXTS } from '../../resources/texts'
+import OperatorSelector from '../components/OperatorSelector'
+import LineNumberSelector from '../components/LineSelector'
+import { SearchContext } from '../../model/pageState'
+import { Gap, GapsList } from '../../model/gaps'
+import { getGapsAsync } from '../../api/gapsService'
 import { Spin } from 'antd'
-import RouteSelector from './components/RouteSelector'
-import { NotFound } from './components/NotFound'
-import { getRoutesAsync } from '../api/gtfsService'
+import RouteSelector from '../components/RouteSelector'
+import { NotFound } from '../components/NotFound'
+import { getRoutesAsync } from '../../api/gtfsService'
 import moment, { Moment } from 'moment'
 import styled from 'styled-components'
 import { useSessionStorage } from 'usehooks-ts'
-import { DateSelector } from './components/DateSelector'
+import { DateSelector } from '../components/DateSelector'
 import { FormControlLabel, Switch } from '@mui/material'
 import Grid from '@mui/material/Unstable_Grid2' // Grid version 2
 import { INPUT_SIZE } from 'src/resources/sizes'
-import DisplayGapsPercentage from './components/DisplayGapsPercentage'
+import DisplayGapsPercentage from '../components/DisplayGapsPercentage'
 
 function formatTime(time: Moment) {
   return time.format(TEXTS.time_format)
