+++ src/pages/realtimeMap/index.tsx
@@ -9,16 +9,16 @@ import getAgencyList, { Agency } from 'src/api/agencyList'
 import useVehicleLocations from 'src/api/useVehicleLocations'
 import { VehicleLocation } from 'src/model/vehicleLocation'
 import './Map.scss'
-import { DateSelector } from './components/DateSelector'
-import MinuteSelector from './components/MinuteSelector'
+import { DateSelector } from '../components/DateSelector'
+import MinuteSelector from '../components/MinuteSelector'
 import Grid from '@mui/material/Unstable_Grid2' // Grid version 2
-import { PageContainer } from './components/PageContainer'
+import { PageContainer } from '../components/PageContainer'
 import { INPUT_SIZE } from 'src/resources/sizes'
-import { Label } from './components/Label'
-import { getColorByHashString } from './dashboard/AllLineschart/OperatorHbarChart/utils'
-import createClusterCustomIcon from './components/utils/customCluster/customCluster'
-import { TimeSelector } from './components/TimeSelector'
-import { busIcon, busIconPath } from './components/utils/BusIcon'
+import { Label } from '../components/Label'
+import { getColorByHashString } from '../dashboard/AllLineschart/OperatorHbarChart/utils'
+import createClusterCustomIcon from '../components/utils/customCluster/customCluster'
+import { TimeSelector } from '../components/TimeSelector'
+import { busIcon, busIconPath } from '../components/utils/BusIcon'
 import { BusToolTip } from 'src/pages/components/MapLayers/BusToolTip'
 
 const { Title } = Typography
