+++ src/pages/singleLineMap/index.tsx
@@ -9,20 +9,20 @@ import OperatorSelector from 'src/pages/components/OperatorSelector'
 import RouteSelector from 'src/pages/components/RouteSelector'
 import { INPUT_SIZE } from 'src/resources/sizes'
 import { TEXTS } from 'src/resources/texts'
-import { SearchContext } from '../model/pageState'
-import { NotFound } from './components/NotFound'
-import { Point } from './RealtimeMapPage'
+import { SearchContext } from '../../model/pageState'
+import { NotFound } from '../components/NotFound'
+import { Point } from '../realtimeMap'
 
 import Grid from '@mui/material/Unstable_Grid2' // Grid version 2
-import './Map.scss'
+import '../Map.scss'
 import getAgencyList, { Agency } from 'src/api/agencyList'
 import { VehicleLocation } from 'src/model/vehicleLocation'
-import { getColorByHashString } from './dashboard/AllLineschart/OperatorHbarChart/utils'
-import { DateSelector } from './components/DateSelector'
+import { getColorByHashString } from '../dashboard/AllLineschart/OperatorHbarChart/utils'
+import { DateSelector } from '../components/DateSelector'
 import { CircularProgress } from '@mui/material'
-import { FilterPositionsByStartTimeSelector } from './components/FilterPositionsByStartTimeSelector'
-import { PageContainer } from './components/PageContainer'
-import { busIcon, busIconPath } from './components/utils/BusIcon'
+import { FilterPositionsByStartTimeSelector } from '../components/FilterPositionsByStartTimeSelector'
+import { PageContainer } from '../components/PageContainer'
+import { busIcon, busIconPath } from '../components/utils/BusIcon'
 import { BusToolTip } from 'src/pages/components/MapLayers/BusToolTip'
 
 interface Path {
